backend/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
pydantic==2.9.2
pymongo==4.10.1
python-dotenv==1.0.1
